Add files via upload
This project focuses on analyzing mobile phone data to uncover patterns and insights related to pricing, customer engagement, and discount trends. Using Python, the Pandas library helps in loading, cleaning, and transforming the data, while the Plotly library is employed to create interactive visualizations that reveal trends in the dataset.